backend/requirements.txt
fastapi==0.115.5
lxml_html_clean==0.4.1
uvicorn[standard]==0.32.1
google-generativeai==0.8.3
trafilatura==2.0.0
pymupdf==1.25.1
httpx==0.28.1
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.3